requirements.txt
fastapi
uvicorn
httpx
python-dotenv
faiss-cpu
sentence-transformers
transformers
torch
PyMuPDF
